Cleanup: remove bcryptaes implementation
BCrypt-AES is not going to be merged into upstream and since
we don't actually need it - I've removed the code behind it.
