+++ docs/docs/concepts/credentials/username-email-password.mdx
@@ -53,26 +53,6 @@ hashers:
     cost: 12
 ```
 
-Similarly, Bcrypt wrapped with AES can be also configured. In addition to
-standard to Bcrypt `cost` parameter (default value is 12) it requires AES `key`
-to be specified. Since we want to use specific AES variant (namely AES-256) the
-`key` parameter must contain 32-bytes long strings (see:
-[docs](https://golang.org/pkg/crypto/aes/#NewCipher)). `key` parameter is
-defined as an array to allow easy key rotation (similarly how secrets are
-implemented) where first key is treated as "current" and others are treated as
-legacy one and are used only to verify hashes that were created using older
-keys.
-
-```yaml title="path/to/my/kratos/config.yml"
-# $ kratos -c path/to/my/kratos/config.yml serve
-hashers:
-  bcryptaes:
-    const: 12
-    key:
-      - ABCDEFGHIJKLMNOPQRSTUVWXYZ123456
-      - abcdefghijklmnopqrstuvwxyz789012
-```
-
 By default, Kratos uses BCrypt algorithm for password hashing. Use the following
 option to use the Argon2id algorithm:
