+++ driver/config/.schema/config.schema.json
@@ -1498,10 +1498,10 @@
       "properties": {
         "algorithm": {
           "title": "Password hashing algorithm",
-          "description": "One of the values: argon2, bcrypt",
+          "description": "One of the values: argon2, bcrypt, legacyfandom",
           "type": "string",
           "default": "bcrypt",
-          "enum": ["argon2", "bcrypt", "bcryptaes", "legacyfandom"]
+          "enum": ["argon2", "bcrypt", "legacyfandom"]
         },
         "argon2": {
           "title": "Configuration for the Argon2id hasher.",
@@ -1567,31 +1567,6 @@
               "minimum": 4,
               "maximum": 31,
               "default": 12
-            }
-          }
-        },
-        "bcryptaes": {
-          "title": "Configuration for the Bcrypt-AES hasher.",
-          "type": "object",
-          "additionalProperties": false,
-          "properties": {
-            "cost": {
-              "title": "Cost of the BCrypt-AES algorithm. Minimum is 4 when --dev flag is used and 12 otherwise.",
-              "type": "integer",
-              "minimum": 4,
-              "maximum": 31,
-              "default": 12
-            },
-            "key": {
-              "type": "array",
-              "title": "Encryption Keys for BCrypt-AES hasher",
-              "description": "The first key in the array is used for encrypting/verifying credentials while all other keys are used to verify older hashes that were encrypted with that old key.",
-              "items": {
-                "type": "string",
-                "minLength": 32,
-                "maxLength": 32
-              },
-              "uniqueItems": true
             }
           }
         },
@@ -1818,35 +1793,6 @@
                 }
               }
             }
-          }
-        }
-      }
-    },
-    {
-      "if": {
-        "properties": {
-          "hashers": {
-            "properties": {
-              "algorithm": {
-                "const": "bcryptaes"
-              }
-            },
-            "required": [ "algorithm" ]
-          }
-        }
-      },
-      "then": {
-        "properties": {
-          "hashers": {
-            "properties": {
-              "bcryptaes": {
-                "required": [
-                  "cost",
-                  "key"
-                ]
-              }
-            },
-            "required": ["bcryptaes"]
           }
         }
       }
