+++ hash/hasher_legacy_fandom.go
@@ -3,8 +3,12 @@ package hash
 import (
 	"bytes"
 	"context"
+	"crypto/aes"
+	"crypto/cipher"
+	"crypto/rand"
 	"encoding/hex"
 	"fmt"
+	"io"
 
 	"github.com/ory/kratos/driver/config"
 	"github.com/pkg/errors"
@@ -22,14 +26,37 @@ type LegacyFandomCryptConfiguration interface {
 	config.Provider
 }
 
+// aes256Encrypt encrypts data using 256-bit AES-GCM.  This both hides the content of
+// the data and provides a check that it hasn't been altered. Output takes the
+// form nonce|ciphertext|tag where '|' indicates concatenation.
+func aes256Encrypt(plaintext []byte, key *[32]byte) (ciphertext []byte, err error) {
+	block, err := aes.NewCipher(key[:])
+	if err != nil {
+		return nil, err
+	}
+
+	gcm, err := cipher.NewGCM(block)
+	if err != nil {
+		return nil, err
+	}
+
+	nonce := make([]byte, gcm.NonceSize())
+	_, err = io.ReadFull(rand.Reader, nonce)
+	if err != nil {
+		return nil, err
+	}
+
+	return gcm.Seal(nonce, nonce, plaintext, nil), nil
+}
+
 func NewHasherLegacyFandom(c LegacyFandomCryptConfiguration) *LegacyFandomCrypt {
 	return &LegacyFandomCrypt{c: c}
 }
 
 func (h *LegacyFandomCrypt) Generate(ctx context.Context, password []byte) ([]byte, error) {
-	cfg := h.c.Config(ctx).HasherLegacyFandom()
-	if len(cfg.Key) == 0 {
-		return nil, NoAESKeyError
+	cfg, err := h.c.Config(ctx).HasherLegacyFandom()
+	if err != nil {
+		return nil, errors.WithStack(err)
 	}
 
 	sh := sha3.New512()
