+++ internal/.kratos.yaml
@@ -34,10 +34,6 @@ hashers:
     dedicated_memory: 1GB
     expected_duration: 500ms
     expected_deviation: 500ms
-  bcryptaes:
-    key:
-      - R3w7Hk5!47$$EqD*3!r5*kEfG97BHvAJ
-      - b3Dx@jtl1yKau0X3$L!5PJQTgjAOH%cJ
   legacyfandom:
     key:
       - QWERTYUIOPASDFGHQWERTYUIOPASDFGH
